+++ src/connect/components/drive-view/drive-view.tsx
@@ -1,6 +1,6 @@
 import { Icon } from '@/powerhouse';
-import { DropItem } from '@/powerhouse/hooks';
-import type { DropEvent } from 'react-aria';
+import type { DropItem } from '@/powerhouse/hooks';
+import type { DragEndEvent, DragStartEvent, DropEvent } from 'react-aria';
 import { Button } from 'react-aria-components';
 import { twMerge } from 'tailwind-merge';
 import {
@@ -25,7 +25,10 @@ export type OnItemOptionsClickHandler<T extends string = DefaultOptionId> = (
 ) => void;
 
 export interface DriveViewProps<T extends string = DefaultOptionId>
-    extends React.HTMLAttributes<HTMLDivElement> {
+    extends Omit<
+        React.HTMLAttributes<HTMLDivElement>,
+        'onDragEnd' | 'onDragStart'
+    > {
     type: DriveType;
     name: string;
     drives: DriveTreeItem<T>[];
@@ -41,7 +44,25 @@ export interface DriveViewProps<T extends string = DefaultOptionId>
         item: TreeItem<T>,
         drive: DriveTreeItem<T>,
     ) => void;
+    onSubmitInput?: (item: TreeItem, drive: DriveTreeItem<T>) => void;
+    onCancelInput?: (item: TreeItem, drive: DriveTreeItem<T>) => void;
     onItemOptionsClick?: OnItemOptionsClickHandler<T>;
+    disableHighlightStyles?: boolean;
+
+    onDropActivate?: (
+        drive: DriveTreeItem<T>,
+        dropTargetItem: TreeItem,
+    ) => void;
+    onDragStart?: (
+        drive: DriveTreeItem<T>,
+        dragItem: TreeItem,
+        event: DragStartEvent,
+    ) => void;
+    onDragEnd?: (
+        drive: DriveTreeItem<T>,
+        dragItem: TreeItem,
+        event: DragEndEvent,
+    ) => void;
 }
 
 export function DriveView<T extends string = DefaultOptionId>(
@@ -54,8 +75,14 @@ export function DriveView<T extends string = DefaultOptionId>(
         drives,
         onDropEvent,
         onItemClick,
+        onSubmitInput,
         onItemOptionsClick,
         defaultItemOptions,
+        onDropActivate,
+        onDragStart,
+        onDragEnd,
+        onCancelInput,
+        disableHighlightStyles,
         ...restProps
     } = props;
     return (
@@ -80,8 +107,20 @@ export function DriveView<T extends string = DefaultOptionId>(
                     <ConnectTreeView<T>
                         key={drive.id}
                         items={drive}
+                        disableHighlightStyles={disableHighlightStyles}
+                        onDragEnd={(...args) => onDragEnd?.(drive, ...args)}
+                        onDragStart={(...args) => onDragStart?.(drive, ...args)}
                         onDropEvent={(...args) => onDropEvent?.(...args, drive)}
                         onItemClick={(...args) => onItemClick?.(...args, drive)}
+                        onDropActivate={(...args) =>
+                            onDropActivate?.(drive, ...args)
+                        }
+                        onCancelInput={(...args) =>
+                            onCancelInput?.(...args, drive)
+                        }
+                        onSubmitInput={(...args) =>
+                            onSubmitInput?.(...args, drive)
+                        }
                         onItemOptionsClick={(...args) =>
                             onItemOptionsClick?.(...args, drive)
                         }