+++ src/connect/components/tree-view/tree-view.stories.tsx
@@ -1,6 +1,7 @@
+import { traverseTree } from '@/connect/utils';
 import type { Meta, StoryObj } from '@storybook/react';
 import { useState } from 'react';
-import { ItemStatus, ItemType, TreeItem } from '../tree-view-item';
+import { ActionType, ItemStatus, ItemType, TreeItem } from '../tree-view-item';
 import { ConnectTreeView, ConnectTreeViewProps } from './tree-view';
 
 const meta = {
@@ -12,6 +13,11 @@ const meta = {
         onDropEvent: { control: { type: 'action' } },
         onItemOptionsClick: { control: { type: 'action' } },
         defaultItemOptions: { control: { type: 'object' } },
+        onSubmitInput: { control: { type: 'action' } },
+        onCancelInput: { control: { type: 'action' } },
+        onDropActivate: { control: { type: 'action' } },
+        onDragStart: { control: { type: 'action' } },
+        onDragEnd: { control: { type: 'action' } },
     },
 } satisfies Meta<typeof ConnectTreeView>;
 
@@ -91,23 +97,19 @@ const treeItem: TreeItem = {
 };
 
 const TreeViewImpl = (args: ConnectTreeViewProps) => {
-    const { onItemClick, items: argItems, ...treeViewProps } = args;
+    const {
+        onItemClick,
+        items: argItems,
+        onItemOptionsClick,
+        onCancelInput,
+        onSubmitInput,
+        onDropActivate,
+        onDragStart,
+        onDragEnd,
+        ...treeViewProps
+    } = args;
     const [items, setItems] = useState(argItems);
-
-    const traverseTree = (
-        item: TreeItem,
-        callback: (item: TreeItem) => TreeItem,
-    ): TreeItem => {
-        const treeItem = callback(item);
-
-        if (treeItem.children) {
-            treeItem.children = treeItem.children.map(child =>
-                traverseTree(child, callback),
-            );
-        }
-
-        return { ...treeItem };
-    };
+    const [disableHighlight, setDisableHighlight] = useState(false);
 
     const onItemClickHandler: ConnectTreeViewProps['onItemClick'] = (
         e,
@@ -117,7 +119,8 @@ const TreeViewImpl = (args: ConnectTreeViewProps) => {
         setItems(prevState => {
             const newTree = traverseTree(prevState, treeItem => {
                 if (treeItem.id === item.id) {
-                    treeItem.isSelected = !treeItem.isSelected;
+                    treeItem.isSelected = true;
+                    treeItem.expanded = !treeItem.expanded;
                 } else {
                     treeItem.isSelected = false;
                 }
@@ -129,11 +132,125 @@ const TreeViewImpl = (args: ConnectTreeViewProps) => {
         });
     };
 
+    const onItemOptionsClickHandler: ConnectTreeViewProps['onItemOptionsClick'] =
+        (item, option) => {
+            onItemOptionsClick?.(item, option);
+
+            if (option === 'rename') {
+                const newTree = traverseTree(items, treeItem => {
+                    if (treeItem.id === item.id) {
+                        treeItem.action = ActionType.Update;
+                    } else {
+                        treeItem.action = undefined;
+                        treeItem.isSelected = false;
+                    }
+
+                    return treeItem;
+                });
+
+                setItems(newTree);
+                return;
+            }
+
+            if (option === 'new-folder') {
+                const newTree = traverseTree(items, treeItem => {
+                    if (treeItem.id === item.id) {
+                        treeItem.expanded = true;
+                        treeItem.isSelected = false;
+                        treeItem.children = treeItem.children || [];
+                        treeItem.children.push({
+                            id: `${treeItem.id}/new-folder`,
+                            label: 'New Folder',
+                            type: ItemType.Folder,
+                            action: ActionType.New,
+                        });
+                    }
+
+                    return treeItem;
+                });
+
+                setItems(newTree);
+            }
+        };
+
+    const onCancelInputHandler: ConnectTreeViewProps['onCancelInput'] =
+        item => {
+            onCancelInput?.(item);
+            const newTree = traverseTree(items, treeItem => {
+                if (treeItem.id === item.id) {
+                    treeItem.action = undefined;
+                }
+
+                return treeItem;
+            });
+
+            setItems(newTree);
+        };
+
+    const onSubmitInputHandler: ConnectTreeViewProps['onSubmitInput'] =
+        item => {
+            onSubmitInput?.(item);
+            const newTree = traverseTree(items, treeItem => {
+                if (treeItem.id === item.id) {
+                    treeItem.action = undefined;
+                    treeItem.label = item.label;
+                    item.id = item.id.replace(
+                        /\/new-folder$/,
+                        `/${item.label}`,
+                    );
+                }
+
+                return treeItem;
+            });
+
+            setItems(newTree);
+        };
+
+    const onDropActivateHandler: ConnectTreeViewProps['onDropActivate'] =
+        item => {
+            onDropActivate?.(item);
+
+            setItems(prevState => {
+                const newTree = traverseTree(prevState, treeItem => {
+                    if (treeItem.id === item.id) {
+                        treeItem.expanded = true;
+                    }
+
+                    return treeItem;
+                });
+
+                return newTree;
+            });
+        };
+
+    const onDragStartHandler: ConnectTreeViewProps['onDragStart'] = (
+        item,
+        event,
+    ) => {
+        onDragStart?.(item, event);
+        setDisableHighlight(true);
+    };
+
+    const onDragEndHandler: ConnectTreeViewProps['onDragEnd'] = (
+        item,
+        event,
+    ) => {
+        onDragEnd?.(item, event);
+        setDisableHighlight(false);
+    };
+
     return (
         <div className="p-10 bg-white">
             <ConnectTreeView
                 items={items}
                 onItemClick={onItemClickHandler}
+                onCancelInput={onCancelInputHandler}
+                onSubmitInput={onSubmitInputHandler}
+                onItemOptionsClick={onItemOptionsClickHandler}
+                onDropActivate={onDropActivateHandler}
+                disableHighlightStyles={disableHighlight}
+                onDragStart={onDragStartHandler}
+                onDragEnd={onDragEndHandler}
                 {...treeViewProps}
             />
         </div>