+++ src/connect/components/tree-view/tree-view.tsx
@@ -1,13 +1,22 @@
 import React from 'react';
 
 import {
+    ActionType,
     ConnectTreeViewItem,
     ConnectTreeViewItemProps,
     TreeItem,
 } from '../tree-view-item';
 
+import {
+    ConnectTreeViewInput,
+    ConnectTreeViewInputProps,
+} from '../tree-view-input';
+
 export interface ConnectTreeViewProps<T extends string = string>
-    extends Omit<React.HTMLAttributes<HTMLElement>, 'onClick'> {
+    extends Omit<
+        React.HTMLAttributes<HTMLElement>,
+        'onClick' | 'onDragStart' | 'onDragEnd'
+    > {
     items: TreeItem<T>;
     onDropEvent?: ConnectTreeViewItemProps<T>['onDropEvent'];
     onItemClick?: (
@@ -16,6 +25,12 @@ export interface ConnectTreeViewProps<T extends string = string>
     ) => void;
     onItemOptionsClick?: ConnectTreeViewItemProps<T>['onOptionsClick'];
     defaultItemOptions?: ConnectTreeViewItemProps<T>['defaultOptions'];
+    onSubmitInput?: ConnectTreeViewInputProps['onSubmit'];
+    onCancelInput?: ConnectTreeViewInputProps['onCancel'];
+    onDropActivate?: ConnectTreeViewItemProps<T>['onDropActivate'];
+    onDragStart?: ConnectTreeViewItemProps<T>['onDragStart'];
+    onDragEnd?: ConnectTreeViewItemProps<T>['onDragEnd'];
+    disableHighlightStyles?: boolean;
 }
 
 export function ConnectTreeView<T extends string = string>(
@@ -27,18 +42,36 @@ export function ConnectTreeView<T extends string = string>(
         onDropEvent,
         defaultItemOptions,
         onItemOptionsClick,
+        onSubmitInput = () => {},
+        onCancelInput = () => {},
         ...elementProps
     } = props;
 
     function renderTreeItems(item: TreeItem<T>, level = 0) {
+        if (
+            item.action === ActionType.New ||
+            item.action === ActionType.Update
+        ) {
+            return (
+                <ConnectTreeViewInput
+                    item={item}
+                    key={item.id}
+                    level={level}
+                    onSubmit={onSubmitInput}
+                    onCancel={onCancelInput}
+                />
+            );
+        }
+
         return (
             <ConnectTreeViewItem<T>
                 item={item}
                 key={item.id}
                 onDropEvent={onDropEvent}
-                onClick={e => onItemClick?.(e, item)}
                 onOptionsClick={onItemOptionsClick}
                 defaultOptions={defaultItemOptions}
+                onClick={e => onItemClick?.(e, item)}
+                disableDropBetween={level === 0 && !item.expanded}
                 {...elementProps}
             >
                 {item.children?.map(item => renderTreeItems(item, level + 1))}