+++ src/powerhouse/components/tree-view-item/tree-view-item.stories.tsx
@@ -56,7 +56,6 @@ const optionsContent = (
 export const Primary: Story = {
     args: {
         label: 'Local Device',
-        initialOpen: true,
         iconName: 'folder-close-fill',
         expandedIconName: 'folder-open-fill',
         secondaryIconName: 'syncing',
@@ -119,7 +118,6 @@ const StyledTreeViewItem: React.FC<TreeViewItemProps> = props => {
 export const WithStyles: Story = {
     args: {
         label: 'Local Device',
-        initialOpen: true,
         iconName: 'folder-close-fill',
         expandedIconName: 'folder-open-fill',
         secondaryIconName: 'syncing',