+++ vite.config.ts
@@ -1,22 +1,25 @@
 import react from '@vitejs/plugin-react';
-import { resolve } from 'path';
+import path, { resolve } from 'path';
 import { defineConfig } from 'vite';
 import dts from 'vite-plugin-dts';
 import { libInjectCss } from 'vite-plugin-lib-inject-css';
-import tsConfigPaths from 'vite-tsconfig-paths';
 import pkg from './package.json';
 
 export default defineConfig(props => {
     return {
         plugins: [
-            tsConfigPaths(),
             react(),
             libInjectCss(),
             dts({
-                include: ['src'],
+                include: ['src/**'],
                 exclude: ['src/**/*.stories.*'],
             }),
         ],
+        resolve: {
+            alias: {
+                '@': path.resolve(__dirname, 'src'),
+            },
+        },
         build: {
             cssCodeSplit: true,
             sourcemap: true,