+++ CHANGELOG.md
@@ -1,3 +1,94 @@
+# [1.0.0-alpha.18](https://github.com/powerhouse-inc/design-system/compare/v1.0.0-alpha.17...v1.0.0-alpha.18) (2023-11-14)
+
+
+### Features
+
+* temp remove icon component ([0f1156c](https://github.com/powerhouse-inc/design-system/commit/0f1156c676dd18ccf26313f143ae1330a72d1f59))
+* temp switch to inline svg ([3beeab9](https://github.com/powerhouse-inc/design-system/commit/3beeab9c563f862a85559c2f854f270239caec37))
+
+# [1.0.0-alpha.17](https://github.com/powerhouse-inc/design-system/compare/v1.0.0-alpha.16...v1.0.0-alpha.17) (2023-11-14)
+
+
+### Bug Fixes
+
+* move icons file to assets directory ([63eeef1](https://github.com/powerhouse-inc/design-system/commit/63eeef10f8aea3354b7901ffe70b1f2e2aba0477))
+* update path for icons file ([5b25ea1](https://github.com/powerhouse-inc/design-system/commit/5b25ea1f9f89523cbcdef5bcda87cdf19172352e))
+
+# [1.0.0-alpha.16](https://github.com/powerhouse-inc/design-system/compare/v1.0.0-alpha.15...v1.0.0-alpha.16) (2023-11-13)
+
+
+### Features
+
+* export breadcrumbs component ([a9c549b](https://github.com/powerhouse-inc/design-system/commit/a9c549b57c55c02a2e8084eb63078ae1b03b7fdd))
+
+# [1.0.0-alpha.15](https://github.com/powerhouse-inc/design-system/compare/v1.0.0-alpha.14...v1.0.0-alpha.15) (2023-11-13)
+
+
+### Features
+
+* fixed dropItem export ([e9319cb](https://github.com/powerhouse-inc/design-system/commit/e9319cb259bdf56a0e9fa374d0f3a1fd71dc3b21))
+
+# [1.0.0-alpha.14](https://github.com/powerhouse-inc/design-system/compare/v1.0.0-alpha.13...v1.0.0-alpha.14) (2023-11-13)
+
+
+### Bug Fixes
+
+* remove stray console log ([a7c1f28](https://github.com/powerhouse-inc/design-system/commit/a7c1f281770482ec90332407829ed44e369c79a4))
+* update prop name ([2994dd8](https://github.com/powerhouse-inc/design-system/commit/2994dd8a746f3c2c6ac22ea5cac996a77a06e520))
+
+
+### Features
+
+* add breadcrumb styles ([93b796b](https://github.com/powerhouse-inc/design-system/commit/93b796b1897e43e4001853ce1c2518d7ebb01033))
+* add breadcrumbs component ([a5b1ddd](https://github.com/powerhouse-inc/design-system/commit/a5b1ddd96fd0131f752403aebf4652846b2536da))
+* add breadcrumbs component ([b1c1d0b](https://github.com/powerhouse-inc/design-system/commit/b1c1d0bc685e7a6dbccd2bf962b8e4344a54574f))
+* add breadcrumbs story ([8a7e593](https://github.com/powerhouse-inc/design-system/commit/8a7e593d49cde456eb5e9e6a6606c998b5e9d187))
+* add comments ([a59237e](https://github.com/powerhouse-inc/design-system/commit/a59237e8ec4c04a3cd229d33effb23785791f7b1))
+* add height to breadcrumbs ([416fc02](https://github.com/powerhouse-inc/design-system/commit/416fc027d1bafc61136573f1eb776d31c4147ffb))
+* add icon component ([e74d306](https://github.com/powerhouse-inc/design-system/commit/e74d30612956705d138a1b6b7eb50e8f7a52d4b0))
+* add new item input logic ([eb7eae7](https://github.com/powerhouse-inc/design-system/commit/eb7eae73ff5e4e5424cee782cad5bebe37ab6ed9))
+* remove redundant props ([d5d181c](https://github.com/powerhouse-inc/design-system/commit/d5d181cad5046f5cef85bb8b6e7c4f093905361a))
+
+# [1.0.0-alpha.13](https://github.com/powerhouse-inc/design-system/compare/v1.0.0-alpha.12...v1.0.0-alpha.13) (2023-11-13)
+
+
+### Features
+
+* 🚀 Added open folder when hover in drag state ([0195eb3](https://github.com/powerhouse-inc/design-system/commit/0195eb3f64e7d2b7bfe4625fd3e6deba350dad87))
+* added missing props to DriveView ([6a862ea](https://github.com/powerhouse-inc/design-system/commit/6a862ea723eea749bcfcef29ce08b0b85daad658))
+* handle highlight styles when dragging an item ([d270a55](https://github.com/powerhouse-inc/design-system/commit/d270a5554572a88ca8ccbb0cf929ac244d18bcac))
+* reserve icon space in tree-view-item and tree-view-input ([ddca274](https://github.com/powerhouse-inc/design-system/commit/ddca274945affb5bc104235384081ad1346701bb))
+
+# [1.0.0-alpha.12](https://github.com/powerhouse-inc/design-system/compare/v1.0.0-alpha.11...v1.0.0-alpha.12) (2023-11-10)
+
+
+### Features
+
+* added onSubmitInput handlers on drive-view ([9da2705](https://github.com/powerhouse-inc/design-system/commit/9da2705ca98402b9f0ac3c718ffb28313f52cee4))
+
+# [1.0.0-alpha.11](https://github.com/powerhouse-inc/design-system/compare/v1.0.0-alpha.10...v1.0.0-alpha.11) (2023-11-10)
+
+
+### Features
+
+* removed react-click-away-listener as peer dependency ([8ed9266](https://github.com/powerhouse-inc/design-system/commit/8ed92661e8aad1c21c6760101686f8c7ce330e8b))
+
+# [1.0.0-alpha.10](https://github.com/powerhouse-inc/design-system/compare/v1.0.0-alpha.9...v1.0.0-alpha.10) (2023-11-10)
+
+
+### Features
+
+* **tree-view-input:** 🚀 Added ConnectTreeViewInput component ([97a82bb](https://github.com/powerhouse-inc/design-system/commit/97a82bb8a8d52144a533fe18147d8a1b979ca3a3))
+* **tree-view-input:** 🚀 Added TreeViewInput Component ([bba0d1e](https://github.com/powerhouse-inc/design-system/commit/bba0d1e1b2946474098c86ef56145e012455d43d))
+
+# [1.0.0-alpha.9](https://github.com/powerhouse-inc/design-system/compare/v1.0.0-alpha.8...v1.0.0-alpha.9) (2023-11-09)
+
+
+### Features
+
+* **drag-and-drop:** 🚀 Added drop between items support ([efb5ea9](https://github.com/powerhouse-inc/design-system/commit/efb5ea960c7023e84bbcd137029e70f44923ff83))
+* **drag-and-drop:** avoid render top and bottom when isn't needed ([50151d1](https://github.com/powerhouse-inc/design-system/commit/50151d1397b5516192692b7e54ec7043b684dac5))
+
 # [1.0.0-alpha.8](https://github.com/powerhouse-inc/design-system/compare/v1.0.0-alpha.7...v1.0.0-alpha.8) (2023-11-08)